Moved SymfonySentinelSession out of the Provider namespace.
